Retreive widget from db into slideshow

+++ backend/src/main/resources/db/migration/V1_7__dashboardRotation.sql
@@ -1,2 +1,2 @@
 ALTER TABLE project ADD COLUMN project_type VARCHAR(15) NOT NULL DEFAULT 'classic';
-ALTER TABLE project ADD COLUMN parent INT DEFAULT NULL;
+ALTER TABLE project ADD COLUMN parent bigint DEFAULT NULL;
