+++ src/main/java/io/suricate/monitoring/model/dto/project/ProjectDto.java
@@ -55,7 +55,7 @@ public class ProjectDto extends AbstractDto {
     /**
      * The project parent
      */
-    private Integer parent;
+    private Long parent;
     /**
      * Number of column in the dashboard
      */
